refactor: fix import issues and errors and refactor categories component ui
