--- app/src/main/java/com/moscow/tudee/presentation/category/CategoriesEvents.kt
@@ -1,5 +1,0 @@
-package com.moscow.tudee.presentation.category
-
-sealed class CategoriesEvents {
-    data class NavigateToTasks(val categoryId: Long) : CategoriesEvents()
-}
