+++ app/src/main/java/com/moscow/tudee/presentation/screen/category/component/AddCategoryBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.moscow.tudee.presentation.category.categoryScreen.component
+package com.moscow.tudee.presentation.screen.category.component
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
@@ -8,7 +8,6 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
