+++ app/src/main/java/com/moscow/tudee/presentation/screen/category/component/DeleteCategoryBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.moscow.tudee.presentation.category.categoryScreen.component
+package com.moscow.tudee.presentation.screen.category.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
